# packages/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vapor RMM API",
    "description": "API for managing devices, commands, and monitoring through the Vapor RMM platform.",
    "version": "1.0.0",
    "contact": {
      "name": "Vapor RMM Support",
      "url": "https://github.com/vapor-rmm/vapor-rmm"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Authenticate user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user_id": { "type": "string" },
                    "email": { "type": "string" },
                    "name": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" },
          "429": { "description": "Too many login attempts" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}